backend/requirements.txt
blinker
cffi
click
cryptography
ecdsa
Flask
Flask-Cors
Flask-Login
Flask-SQLAlchemy
greenlet
itsdangerous
Jinja2
MarkupSafe
mysql-connector-python
pyasn1
pycparser
PyJWT
PyMySQL
python-jose
rsa
six
SQLAlchemy
typing_extensions
Werkzeug
python-dotenv
